Removed testing decoder for old Dickert
