+++ src/CMakeLists.txt
@@ -85,7 +85,6 @@ add_library(r_433 STATIC
     devices/current_cost.c
     devices/danfoss.c
     devices/dickert_mahs.c
-    devices/dickert_pwm.c
     devices/digitech_xc0324.c
     devices/directv.c
     devices/dish_remote_6_3.c
